gemini_api/requirements.txt
django 
djangorestframework 
google-generativeai 
python-dotenv 
requests